Finishing delete post dialog

+++ src/app/posts/post-list/post-list.component.ts
@@ -1,10 +1,12 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy, Input } from '@angular/core';
 import { Post } from '../models/post.model';
 import { PostsService } from '../posts.service';
 import {  Subject } from 'rxjs';
 import { PageEvent } from '@angular/material/paginator';
 import { UsersService } from 'src/app/auth/services/users.service';
 import { takeUntil } from 'rxjs/operators';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent } from './dialog/delete-dialog.component';
 
 @Component({
   selector: 'app-post-list',
@@ -21,8 +23,9 @@ export class PostListComponent implements OnInit, OnDestroy {
   posts: Post[] = [];
   destroy$ = new Subject<boolean>();
 
-
-  constructor(private postsService: PostsService, private usersService: UsersService) { }
+  constructor(private postsService: PostsService,
+              private usersService: UsersService,
+              private _dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.postsService.getPostsLoading()
@@ -60,12 +63,17 @@ export class PostListComponent implements OnInit, OnDestroy {
   }
 
   onDelete(postId: string) {
-    this.postsService.deletePost(postId)
-      .pipe(takeUntil(this.destroy$))
-      .subscribe(() => {
-        this.postsService.getPosts(this.postsPerPage, this.currentPage);
-        this.isLoading = false;
-      });
+    const dialogRef = this._dialog.open(DeleteDialogComponent);
+    dialogRef.afterClosed().subscribe(wantsToDelete => {
+      if (wantsToDelete) {
+        this.postsService.deletePost(postId)
+        .pipe(takeUntil(this.destroy$))
+        .subscribe(() => {
+          this.postsService.getPosts(this.postsPerPage, this.currentPage);
+          this.isLoading = false;
+        });
+      }
+    });
   }
 
   ngOnDestroy() {
